refactor: 기물 초기 위치 테스트를 PiecesTest로 이동
